Move error messages from `config-utils` to their own file
