+++ lib/init-action.js
@@ -87045,6 +87045,46 @@ async function getDiffInformedAnalysisBranches(codeql, features, logger) {
   return branches;
 }
 
+// src/error-messages.ts
+var PACKS_PROPERTY = "packs";
+function getConfigFileOutsideWorkspaceErrorMessage(configFile) {
+  return `The configuration file "${configFile}" is outside of the workspace`;
+}
+function getConfigFileDoesNotExistErrorMessage(configFile) {
+  return `The configuration file "${configFile}" does not exist`;
+}
+function getConfigFileRepoFormatInvalidMessage(configFile) {
+  let error2 = `The configuration file "${configFile}" is not a supported remote file reference.`;
+  error2 += " Expected format <owner>/<repository>/<file-path>@<ref>";
+  return error2;
+}
+function getConfigFileFormatInvalidMessage(configFile) {
+  return `The configuration file "${configFile}" could not be read`;
+}
+function getConfigFileDirectoryGivenMessage(configFile) {
+  return `The configuration file "${configFile}" looks like a directory, not a file`;
+}
+function getConfigFilePropertyError(configFile, property, error2) {
+  if (configFile === void 0) {
+    return `The workflow property "${property}" is invalid: ${error2}`;
+  } else {
+    return `The configuration file "${configFile}" is invalid: property "${property}" ${error2}`;
+  }
+}
+function getPacksStrInvalid(packStr, configFile) {
+  return configFile ? getConfigFilePropertyError(
+    configFile,
+    PACKS_PROPERTY,
+    `"${packStr}" is not a valid pack`
+  ) : `"${packStr}" is not a valid pack`;
+}
+function getNoLanguagesError() {
+  return "Did not detect any languages to analyze. Please update input in workflow or check that GitHub detects the correct languages in your repository.";
+}
+function getUnknownLanguagesError(languages) {
+  return `Did not recognize the following languages: ${languages.join(", ")}`;
+}
+
 // src/languages.ts
 var KnownLanguage = /* @__PURE__ */ ((KnownLanguage2) => {
   KnownLanguage2["actions"] = "actions";
@@ -87170,44 +87210,6 @@ async function cachePrefix(codeql, language) {
 }
 
 // src/config-utils.ts
-var PACKS_PROPERTY = "packs";
-function getPacksStrInvalid(packStr, configFile) {
-  return configFile ? getConfigFilePropertyError(
-    configFile,
-    PACKS_PROPERTY,
-    `"${packStr}" is not a valid pack`
-  ) : `"${packStr}" is not a valid pack`;
-}
-function getConfigFileOutsideWorkspaceErrorMessage(configFile) {
-  return `The configuration file "${configFile}" is outside of the workspace`;
-}
-function getConfigFileDoesNotExistErrorMessage(configFile) {
-  return `The configuration file "${configFile}" does not exist`;
-}
-function getConfigFileRepoFormatInvalidMessage(configFile) {
-  let error2 = `The configuration file "${configFile}" is not a supported remote file reference.`;
-  error2 += " Expected format <owner>/<repository>/<file-path>@<ref>";
-  return error2;
-}
-function getConfigFileFormatInvalidMessage(configFile) {
-  return `The configuration file "${configFile}" could not be read`;
-}
-function getConfigFileDirectoryGivenMessage(configFile) {
-  return `The configuration file "${configFile}" looks like a directory, not a file`;
-}
-function getConfigFilePropertyError(configFile, property, error2) {
-  if (configFile === void 0) {
-    return `The workflow property "${property}" is invalid: ${error2}`;
-  } else {
-    return `The configuration file "${configFile}" is invalid: property "${property}" ${error2}`;
-  }
-}
-function getNoLanguagesError() {
-  return "Did not detect any languages to analyze. Please update input in workflow or check that GitHub detects the correct languages in your repository.";
-}
-function getUnknownLanguagesError(languages) {
-  return `Did not recognize the following languages: ${languages.join(", ")}`;
-}
 async function getSupportedLanguageMap(codeql, features, logger) {
   const resolveSupportedLanguagesUsingCli = await features.getValue(
     "resolve_supported_languages_using_cli" /* ResolveSupportedLanguagesUsingCli */,
@@ -87279,7 +87281,9 @@ async function getLanguages(codeql, languagesInput, repository, sourceRoot, feat
   }
   const languages = Array.from(languagesSet);
   if (!autodetected && unknownLanguages.length > 0) {
-    throw new ConfigurationError(getUnknownLanguagesError(unknownLanguages));
+    throw new ConfigurationError(
+      getUnknownLanguagesError(unknownLanguages)
+    );
   }
   if (languages.length === 0) {
     throw new ConfigurationError(getNoLanguagesError());
@@ -87766,7 +87770,9 @@ async function getRemoteConfig(configFile, apiDetails) {
       getConfigFileDirectoryGivenMessage(configFile)
     );
   } else {
-    throw new ConfigurationError(getConfigFileFormatInvalidMessage(configFile));
+    throw new ConfigurationError(
+      getConfigFileFormatInvalidMessage(configFile)
+    );
   }
   return load(
     Buffer.from(fileContents, "base64").toString("binary")
